+++ src/kernel_sidecar/handlers.py
@@ -1,3 +1,5 @@
+import collections
+
 from kernel_sidecar.models import messages
 
 
@@ -25,3 +27,23 @@ class Handler:
         Called when a message is delegated to this Action but no handler is defined for the msg_type
         """
         pass
+
+
+class DebugHandler(Handler):
+    """
+    Useful for testing and debugging. Example:
+
+    handler = DebugHandler()
+    action = kernel.kernel_info_request(handlers=[handler])
+    await action
+    assert handler.counts == {"status": 2, "kernel_info_reply": 1}
+    """
+
+    def __init__(self):
+        self.counts = collections.defaultdict(int)
+        self.last_msg_by_type = collections.defaultdict(messages.Message)
+
+    async def unhandled_message(self, msg: messages.Message):
+        # effectively a catch-all for every message type since no other handlers are defined
+        self.counts[msg.msg_type] += 1
+        self.last_msg_by_type[msg.msg_type] = msg